requirements for "!clear"

+++ commands/clear.js
@@ -4,16 +4,24 @@ module.exports =
   description: 'clear messages',
   async execute(message, args)
   {
-    if((!args[0]) || (isNaN(args[0])) || (args[0] > 1000) || (args[0] < 0))
+    const modoRole = message.guild.roles.cache.find(role => role.name ==='modo');
+    if(message.member.roles.cache.has('825520415027363860'))
     {
-      return message.reply('Vous devez rentrer une valeur entre 0 et 1000');
+      if((!args[0]) || (isNaN(args[0])) || (args[0] > 1000) || (args[0] < 0))
+      {
+        return message.reply('Vous devez rentrer une valeur entre 0 et 1000');
+      }
+      args[0] = (parseInt(args[0])+ 1 ).toString();
+      await message.channel.messages.fetch({limit: args[0]}).then(messages =>
+      {
+        message.channel.bulkDelete(messages);
+      });
+      console.log('cleared');
+      console.log(args[0]);
     }
-    args[0] = (parseInt(args[0])+ 1 ).toString();
-    await message.channel.messages.fetch({limit: args[0]}).then(messages =>
+    else
     {
-      message.channel.bulkDelete(messages);
-    });
-    console.log('cleared');
-    console.log(args[0]);
+      return message.reply("vous n'avez pas l'autorisation de faire ceci")
+    }
   }
 }
